--- a/bin/pycheck
+++ b/bin/pycheck
@@ -33,8 +33,4 @@ try bin/pyactivate -m ruff --extend-exclude=misc/dbt-materialize .
 # We need to maintain compatibility with older Python versions for this
 try bin/pyactivate -m ruff --target-version=py38 misc/dbt-materialize
 
-if [[ ! "${MZDEV_NO_PYTHON_DOCTEST:-}" ]]; then
-  try bin/pyactivate -m pytest -qq --doctest-modules misc/python
-fi
-
 try_status_report