--- a/ci/test/lint-main/checks/check-generation.sh
+++ b/ci/test/lint-main/checks/check-generation.sh
@@ -1,0 +1,23 @@
+#!/usr/bin/env bash
+
+# Copyright Materialize, Inc. and contributors. All rights reserved.
+#
+# Use of this software is governed by the Business Source License
+# included in the LICENSE file at the root of this repository.
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the Apache License, Version 2.0.
+#
+# lint — complains about misformatted files and other problems.
+
+set -euo pipefail
+
+cd "$(dirname "$0")/../../../.."
+
+. misc/shlib/shlib.bash
+
+try bin/gen-completion check
+try bin/gen-lints
+
+try_status_report