--- a/misc/python/materialize/lint/lint.py
+++ b/misc/python/materialize/lint/lint.py
@@ -6,7 +6,7 @@
 # As of the Change Date specified in that file, in accordance with
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
-
+import argparse
 import os
 import subprocess
 import threading
@@ -20,83 +20,104 @@ MAIN_CHECKS_PATH = MAIN_PATH / "checks"
 CHECK_BEFORE_PATH = MAIN_PATH / "before"
 CHECK_AFTER_PATH = MAIN_PATH / "after"
 
-PRINT_DURATION = False
-OMIT_OUTPUT_OF_SUCCESSFUL_CHECK = True
+
+def parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        prog="lint",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description="Lint the code",
+    )
+    parser.add_argument("--print-duration", action="store_true")
+    parser.add_argument("--verbose", action="store_true")
+    return parser.parse_args()
 
 
 def main() -> int:
-    failed_checks = run_and_validate_if_no_previous_failures(
-        CHECK_BEFORE_PATH, previous_failures=[]
-    )
-    failed_checks = run_and_validate_if_no_previous_failures(
-        MAIN_CHECKS_PATH, previous_failures=failed_checks
-    )
-    failed_checks = run_and_validate_if_no_previous_failures(
-        CHECK_AFTER_PATH, previous_failures=failed_checks
-    )
+    args = parse_args()
+    print_duration = args.print_duration
+    verbose_output = args.verbose
 
-    success = len(failed_checks) == 0
-
-    print("+++ Linting result")
-    print("All checks successful." if success else f"Checks failed: {failed_checks}")
-
-    return 0 if success else 1
+    manager = LintManager(print_duration, verbose_output)
+    return manager.run()
 
 
-def is_ignore_file(path: Path) -> bool:
-    return os.path.isdir(path)
+class LintManager:
+    def __init__(self, print_duration: bool, verbose_output: bool):
+        self.print_duration = print_duration
+        self.verbose_output = verbose_output
 
-
-def run_and_validate_if_no_previous_failures(
-    checks_path: Path, previous_failures: list[str]
-) -> list[str]:
-    if len(previous_failures) > 0:
-        print(f"--- Skipping checks in '{checks_path}' due to previous failures")
-        return previous_failures
-    else:
-        return run_and_validate(checks_path)
-
-
-def run_and_validate(checks_path: Path) -> list[str]:
-    """
-    Runs checks in the given directory and validates their outcome.
-    :return: names of failed checks
-    """
-    print(f"--- Running checks in {checks_path}")
-
-    paths = os.listdir(checks_path)
-
-    threads = []
-
-    for lint_file in paths:
-        if is_ignore_file(checks_path / lint_file):
-            continue
-
-        thread = LintingThread(checks_path, lint_file)
-        thread.start()
-        threads.append(thread)
-
-    for thread in threads:
-        thread.join()
-
-    failed_checks = []
-
-    for thread in threads:
-        formatted_duration = (
-            f" in {thread.duration.total_seconds():.2f}s" if PRINT_DURATION else ""
+    def run(self) -> int:
+        failed_checks = self.run_and_validate_if_no_previous_failures(
+            CHECK_BEFORE_PATH, previous_failures=[]
         )
-        if thread.success:
-            print(f"--- {thread.name} (SUCCEEDED{formatted_duration})")
+        failed_checks = self.run_and_validate_if_no_previous_failures(
+            MAIN_CHECKS_PATH, previous_failures=failed_checks
+        )
+        failed_checks = self.run_and_validate_if_no_previous_failures(
+            CHECK_AFTER_PATH, previous_failures=failed_checks
+        )
+
+        success = len(failed_checks) == 0
+
+        print("+++ Linting result")
+        print(
+            "All checks successful." if success else f"Checks failed: {failed_checks}"
+        )
+
+        return 0 if success else 1
+
+    def is_ignore_file(self, path: Path) -> bool:
+        return os.path.isdir(path)
+
+    def run_and_validate_if_no_previous_failures(
+        self, checks_path: Path, previous_failures: list[str]
+    ) -> list[str]:
+        if len(previous_failures) > 0:
+            print(f"--- Skipping checks in '{checks_path}' due to previous failures")
+            return previous_failures
         else:
-            print(f"+++ {thread.name} (FAILED{formatted_duration})")
-            failed_checks.append(thread.name)
+            return self.run_and_validate(checks_path)
 
-        if thread.has_output() and (
-            not thread.success or not OMIT_OUTPUT_OF_SUCCESSFUL_CHECK
-        ):
-            print(thread.output)
+    def run_and_validate(self, checks_path: Path) -> list[str]:
+        """
+        Runs checks in the given directory and validates their outcome.
+        :return: names of failed checks
+        """
+        print(f"--- Running checks in {checks_path}")
 
-    return failed_checks
+        paths = os.listdir(checks_path)
+
+        threads = []
+
+        for lint_file in paths:
+            if self.is_ignore_file(checks_path / lint_file):
+                continue
+
+            thread = LintingThread(checks_path, lint_file)
+            thread.start()
+            threads.append(thread)
+
+        for thread in threads:
+            thread.join()
+
+        failed_checks = []
+
+        for thread in threads:
+            formatted_duration = (
+                f" in {thread.duration.total_seconds():.2f}s"
+                if self.print_duration
+                else ""
+            )
+            if thread.success:
+                print(f"--- {thread.name} (SUCCEEDED{formatted_duration})")
+            else:
+                print(f"+++ {thread.name} (FAILED{formatted_duration})")
+                failed_checks.append(thread.name)
+
+            if thread.has_output() and (not thread.success or self.verbose_output):
+                print(thread.output)
+
+        return failed_checks
 
 
 class LintingThread(threading.Thread):