This parallelizes linting steps. 

Out of scope:
* Make sure that the same checks are executed locally and in CI. This will be done in a follow-up PR.

### Duration Measurements
* Before: <s>23 sec</s> 26 sec (23 sec with `MZDEV_NO_PYTHON_DOCTEST=true`)
* After: 9 sec
* **Speed-up: 2.9**

### Individual Checks
Checks in `checks` will only be run if `before` is successful; checks in `after` will only be run if `checks` are successful.

```
├── after
│   └── check-no-diff.sh
├── before
│   └── can-lint.sh
└── checks
    ├── check-cargo.sh
    ├── check-copyright.sh
    ├── check-generation.sh
    ├── check-new-line.sh
    ├── check-python-docs.sh
    ├── check-python-files.sh
    ├── check-rust-files.sh
    ├── check-rust-version.sh
    └── check-shell-files.sh
```

### Risks and Side Effects
* Coloring seems to get lost for some steps (but not everywhere?), e.g., for `git diff`.

### Exemplary Output

#### Successful run
```
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/before
--- can-lint.sh (SUCCEEDED)
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/checks
--- check-shell-files.sh (SUCCEEDED)
1/1 commands passed
--- check-copyright.sh (SUCCEEDED)
1/1 commands passed
--- check-python-files.sh (SUCCEEDED)
0 errors, 0 warnings, 0 informations 
3/3 commands passed
3/3 commands passed
2/2 commands passed
--- check-rust-files.sh (SUCCEEDED)
1/1 commands passed
--- check-cargo.sh (SUCCEEDED)
1/1 commands passed
--- check-rust-version.sh (SUCCEEDED)
1/1 commands passed
--- check-new-line.sh (SUCCEEDED)
2/2 commands passed
--- check-generation.sh (SUCCEEDED)
2/2 commands passed
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/after
--- check-no-diff.sh (SUCCEEDED)
1/1 commands passed
+++ Linting result
All checks successful.
```

#### Successful run including duration information (not enabled by default)
```
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/before
--- can-lint.sh (SUCCEEDED in 0.02s)
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/checks
--- check-shell-files.sh (SUCCEEDED in 2.40s)
1/1 commands passed
--- check-copyright.sh (SUCCEEDED in 5.43s)
1/1 commands passed
--- check-python-files.sh (SUCCEEDED in 8.69s)
0 errors, 0 warnings, 0 informations 
3/3 commands passed
3/3 commands passed
2/2 commands passed
--- check-rust-files.sh (SUCCEEDED in 4.21s)
1/1 commands passed
--- check-cargo.sh (SUCCEEDED in 0.21s)
1/1 commands passed
--- check-rust-version.sh (SUCCEEDED in 0.03s)
1/1 commands passed
--- check-new-line.sh (SUCCEEDED in 6.91s)
2/2 commands passed
--- check-generation.sh (SUCCEEDED in 0.79s)
2/2 commands passed
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/after
--- check-no-diff.sh (SUCCEEDED in 0.05s)
1/1 commands passed
+++ Linting result
All checks successful.
```

#### Run with a failed check (`after` part is skipped)
```
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/before
--- can-lint.sh (SUCCEEDED)
--- Running checks in /Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/checks
--- check-shell-files.sh (SUCCEEDED)
1/1 commands passed
--- check-copyright.sh (SUCCEEDED)
1/1 commands passed
+++ check-python-files.sh (FAILED)
/Users/nrainer/Workspaces/mz-repo/misc/python/materialize/lint/lint.py
  /Users/nrainer/Workspaces/mz-repo/misc/python/materialize/lint/lint.py:131:16 - error: Expression of type "bool" cannot be assigned to return type "str"
    "bool" is incompatible with "str" (reportGeneralTypeIssues)
1 error, 0 warnings, 0 informations 
^^^ 🚨 Failed: npx --yes pyright@1.1.323 --warnings ci misc/python misc/dbt-materialize/mzcompose.py misc/monitoring/mzcompose.py misc/python/materialize/cli/mzcompose.py misc/rqg/mzcompose.py test/
2/3 commands passed
^^^ 🚨 Failed: bin/pycheck
3/3 commands passed
1/2 commands passed
--- check-rust-files.sh (SUCCEEDED)
1/1 commands passed
--- check-cargo.sh (SUCCEEDED)
1/1 commands passed
--- check-rust-version.sh (SUCCEEDED)
1/1 commands passed
--- check-new-line.sh (SUCCEEDED)
2/2 commands passed
--- check-generation.sh (SUCCEEDED)
2/2 commands passed
--- Skipping checks in '/Users/nrainer/Workspaces/mz-repo/ci/test/lint-main/after' due to previous failures
+++ Linting result
Checks failed: ['check-python-files.sh']
```