@nrainer-materialize 2023-12-12
> One thing that this PR misses is doing the `bin/lint` step of `bin/pre-push` in parallel with the first `cargo` command in there. (Later `cargo` commands can't be in parallel, sadly.) Not sure this is a big deal, savings are small.

Thanks, I wasn't really aware of `bin/pre-push`. I will include this in a follow-up in which I will also make sure that CI runs the same set of checks (which is currently not the case).